Trace the call sites that make malloc/new operations. (#11)
malloc/new operations introduce indeterminism to execution time as kernel need to
handle dynamic memory allocation. This patch reveals the call sites that made
the most malloc/new. An additional python script is also created to reveal
detailed filenames and line numbers of the callers if such information is available.

Co-authored-by: chtseng <chtseng@itri.org.tw>